requirements.txt
aspose_words==23.7.0
openai==0.27.8
pandas==2.0.2
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.25.0
